# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "汽车健康管家平台 S0 API 草案",
    "version": "0.1.0",
    "description": "43 个来源核心接口加 S0 支撑接口。未明确的业务字段以 S0_BUSINESS_CONTRACT.md 为准；未实现的路径不代表已可调用。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/api/payments/create": {
      "post": {
        "summary": "本人支付发起",
        "operationId": "post_api_payments_create",
        "tags": [
          "payments"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreatedPayment"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "409": {
            "description": "订单/支付状态或事件冲突"
          },
          "503": {
            "description": "渠道未配置或数据库/事务暂不可用"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "order_id",
                  "channel"
                ],
                "properties": {
                  "order_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "channel": {
                    "type": "string",
                    "enum": [
                      "LOCAL_TEST",
                      "WECHAT"
                    ]
                  }
                }
              }
            }
          }
        },
        "x-roles": "OWNER",
        "x-implementation-status": "payment-foundation-local-test-implemented",
        "description": "正式微信未配置返回503；LOCAL_TEST默认关闭且仅用于隔离环境，不代表真实扣款。详见 PAYMENT_FOUNDATION.md。"
      }
    },
...
    "/api/payments/{id}": {
      "get": {
        "summary": "本人支付状态查询",
        "operationId": "get_api_payments_id",
        "tags": [
          "payments"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaymentDetail"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "409": {
            "description": "订单/支付状态或事件冲突"
          },
          "503": {
            "description": "渠道未配置或数据库/事务暂不可用"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "OWNER",
        "x-implementation-status": "payment-foundation-local-test-implemented",
        "description": "正式微信未配置返回503；LOCAL_TEST默认关闭且仅用于隔离环境，不代表真实扣款。详见 PAYMENT_FOUNDATION.md。"
      }
    },
...
    "/api/payments/callback/LOCAL_TEST": {
      "post": {
        "summary": "隔离测试渠道签名通知",
        "operationId": "post_api_payments_callback_LOCAL_TEST",
        "tags": [
          "payments"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "X-Test-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Test-Nonce",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Test-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已提交或同事件已提交",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code"
                  ],
                  "properties": {
                    "code": {
                      "const": "SUCCESS"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "通知参数/金额等不匹配"
          },
          "401": {
            "description": "无效/过期签名"
          },
          "404": {
            "description": "测试渠道关闭或支付不存在"
          },
          "409": {
            "description": "事件异文或流水号冲突"
          },
          "503": {
            "description": "事务失败，重试"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "event_id",
                  "payment_id",
                  "channel_payment_no",
                  "status",
                  "amount",
                  "currency",
                  "order_no",
                  "occurred_at"
                ],
                "properties": {
                  "event_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "payment_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "channel_payment_no": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9_-]{1,128}$"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "SUCCEEDED",
                      "FAILED"
                    ]
                  },
                  "amount": {
                    "type": "string",
                    "pattern": "^(?:0|[1-9][0-9]{0,7})\\.[0-9]{2}$"
                  },
                  "currency": {
                    "type": "string",
                    "const": "CNY"
                  },
                  "order_no": {
                    "type": "string"
                  },
                  "occurred_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "x-roles": "LOCAL_TEST signed channel",
        "x-implementation-status": "payment-foundation-local-test-implemented",
        "description": "默认关闭时404；启用须隔离profile/开关/密钥。正文最多16KiB且拒绝重复键；签名及规范化事件摘要规则见 PAYMENT_FOUNDATION.md，不能作为微信验签协议。"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "PaymentDetail": {
        "type": "object",
        "required": [
          "payment_id",
          "order_id",
          "payment_no",
          "channel",
          "test_mode",
          "status",
          "amount",
          "currency",
          "requires_review",
          "expires_at",
          "paid_at"
        ],
        "properties": {
          "payment_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "order_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "payment_no": {
            "type": [
              "string",
              "null"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "LOCAL_TEST",
              "WECHAT"
            ]
          },
          "test_mode": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "SUCCEEDED",
              "FAILED",
              "CLOSED"
            ]
          },
          "amount": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,7})\\.[0-9]{2}$"
          },
          "currency": {
            "type": "string",
            "const": "CNY"
          },
          "requires_review": {
            "type": "boolean"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "CreatedPayment": {
        "type": "object",
        "required": [
          "payment_id",
          "order_id",
          "payment_no",
          "channel",
          "test_mode",
          "status",
          "amount",
          "currency",
          "requires_review",
          "expires_at",
          "paid_at",
          "channel_payload"
        ],
        "properties": {
          "payment_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "order_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "payment_no": {
            "type": [
              "string",
              "null"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "LOCAL_TEST",
              "WECHAT"
            ]
          },
          "test_mode": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PENDING",
              "SUCCEEDED",
              "FAILED",
              "CLOSED"
            ]
          },
          "amount": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,7})\\.[0-9]{2}$"
          },
          "currency": {
            "type": "string",
            "const": "CNY"
          },
          "requires_review": {
            "type": "boolean"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "channel_payload": {
            "type": "object"
          }
        }
      },
...
    }
  }
}